Update with Ramon's PR review.
