+++ itachallenge-challenge/src/main/resources/application.yml
@@ -68,4 +68,4 @@ messages:
 
 user:
   service:
-    url: http://localhost:8764/itachallenge/api/v1/user
+    url: http://apisix-gateway:9080 # IN PRODUCTION
